requirements.txt
openai
chromadb
pandas
python-dotenv